routeに認証系を追加

+++ src/laravel/routes/api.php
@@ -7,6 +7,7 @@ use App\Http\Controllers\TestAPIController;
 use App\Http\Controllers\PreRegisterUserAPIController;
 use App\Http\Controllers\VerifyTokenController;
 use App\Http\Controllers\UserAPIController;
+use App\Http\Controllers\AuthAPIController;
 
 
 /*
@@ -32,9 +33,14 @@ Route::post('/user/pre_register', [PreRegisterUserAPIController::class, 'store']
 // 仮会員登録のtoken確認
 Route::get('/verify/{token}', [VerifyTokenController::class, 'index']);
 
+// 認証
+Route::post('/login', [AuthAPIController::class, 'login']);
+Route::post('/logout', [AuthAPIController::class, 'logout']);
+
 // 会員情報
-// TODO tokenによる認証周りはログイン機能を実装してからやる
 Route::post('/user', [UserAPIController::class, 'store']);
-Route::get('/user/{user_id}', [UserAPIController::class, 'show'])->middleware('loginUserCheck');
-Route::put('/user/{user_id}', [UserAPIController::class, 'update'])->middleware('loginUserCheck');
-Route::delete('/user/{user_id}', [UserAPIController::class, 'destroy'])->middleware('loginUserCheck');
+Route::get('/user/{user_id}', [UserAPIController::class, 'show'])->middleware('auth:sanctum','loginUserCheck');
+Route::put('/user/{user_id}', [UserAPIController::class, 'update'])->middleware('auth:sanctum','loginUserCheck');
+Route::delete('/user/{user_id}', [UserAPIController::class, 'destroy'])->middleware('auth:sanctum','loginUserCheck');
+
+
